% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_plot.R
\name{box_plot}
\alias{box_plot}
\title{Create box plot}
\usage{
box_plot(lipid_data, title = "")
}
\arguments{
\item{lipid_data}{tibble with all the lipid data and test data}

\item{title}{title of the plot}
}
\value{
plotly object
}
\description{
Create box plot.
}
\author{
Rico Derks
}
