% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_plot.R
\name{bubble_plot}
\alias{bubble_plot}
\title{Create bubble plot}
\usage{
bubble_plot(lipid_data, pattern = NULL, title = "")
}
\arguments{
\item{lipid_data}{tibble with all the lipid data and test data}

\item{pattern}{regular expression used to select the correct lipids}

\item{title}{title of the plot}
}
\value{
ggplot object
}
\description{
Create a bubble plot.
}
\details{
This bubble plot function is used to create the bubble plots in the
    report.
}
\author{
Rico Derks
}
