% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubblePlotServer.R
\name{bubblePlotServer}
\alias{bubblePlotServer}
\title{Create bubble plot server part}
\usage{
bubblePlotServer(id, data)
}
\arguments{
\item{id}{id of the server part}

\item{data}{the data for the server part}
}
\value{
a part of the server
}
\description{
Create the server part for the bubble plot
}
\author{
Rico Derks
}
