% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_blank_ratio.R
\name{calc_blank_ratio}
\alias{calc_blank_ratio}
\title{Calculate the sample to average blank ratio}
\usage{
calc_blank_ratio(lipid_data)
}
\arguments{
\item{lipid_data}{tibble in tidy format}
}
\value{
Returns a tibble in tidy with RSD values for each lipid.
}
\description{
Calculate for each lipid the ratio between the sample and the
    average of all blanks.
}
\author{
Rico Derks
}
