% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rsd.R
\name{calc_rsd}
\alias{calc_rsd}
\title{Calculate RSD values of all lipids}
\usage{
calc_rsd(lipid_data)
}
\arguments{
\item{lipid_data}{tibble in tidy format}
}
\value{
Returns a tibble in tidy with RSD values for each lipid.
}
\description{
Calculate RSD values of all lipids within all QCpool samples..
}
\author{
Rico Derks
}
