% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_up.R
\name{clean_up}
\alias{clean_up}
\title{Clean up tibble from MS-DIAL}
\usage{
clean_up(lipid_data)
}
\arguments{
\item{lipid_data}{The tibble.}
}
\value{
Returns a tibble
}
\description{
Clean up the columns and column names of the tibble after
    reading the MS-DIAL result files.
}
\author{
Rico Derks
}
