% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_samples_heatmap.R
\name{compare_samples_heatmap}
\alias{compare_samples_heatmap}
\title{Create correlation heatmap of all samples}
\usage{
compare_samples_heatmap(
  lipid_data,
  cent_scale,
  z,
  clust = FALSE,
  sample_group = NULL
)
}
\arguments{
\item{lipid_data}{tibble with all the lipid data}

\item{cent_scale}{logical, apply center and scaling}

\item{z}{what to show as intensity of the heatmap}

\item{clust}{apply clustering yes/no, default is no}

\item{sample_group}{dataframe with grouping information}
}
\value{
plotly object
}
\description{
Create a correlation heatmap of all samples..
}
\author{
Rico Derks
}
