% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_mwtest.R
\name{do_mwtest}
\alias{do_mwtest}
\title{Do a Mann-Whitney U test on all lipids}
\usage{
do_mwtest(lipid_data)
}
\arguments{
\item{lipid_data}{tibble in tidy format, already nested}
}
\value{
a tibble with the results and the following columns
    model_test contains the model information for each lipid (nested)
    pvalue contains the uncorrected pvalue
    pvalue_cor contains the corrected pvalue
    fc the fold change estimate1/estimate2
}
\description{
Do a Mann-Whitney U test on all lipids
}
\details{
A Mann-Whitney U test will be done for each lipid. Also the p-value will be corrected
    for multiple testing.
}
\author{
Rico Derks
}
