% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_pca.R
\name{do_pca}
\alias{do_pca}
\title{Do PCA analysis}
\usage{
do_pca(lipid_data)
}
\arguments{
\item{lipid_data}{tibble in tidy format}
}
\value{
a list with scores, loadings and explained variance
}
\description{
Do a quick PCA analysis.
}
\author{
Rico Derks
}
