% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_ttest.R
\name{do_ttest}
\alias{do_ttest}
\title{Do a t-test on all lipids}
\usage{
do_ttest(lipid_data)
}
\arguments{
\item{lipid_data}{tibble in tidy format, already nested}
}
\value{
a tibble with the results and the following columns
    model_test contains the model information for each lipid (nested)
    pvalue contains the uncorrected pvalue
    pvalue_cor contains the corrected pvalue
    fc the fold change estimate1/estimate2
}
\description{
Do a t-test on all lipids.
}
\details{
A t-test will be done for each lipid. Also the p-value will be corrected
    for multiple testing.
}
\author{
Rico Derks
}
