% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launchApp.R
\name{launchApp}
\alias{launchApp}
\title{launches the shiny app}
\usage{
launchApp()
}
\value{
Shiny application object.
}
\description{
This starts the app locally.
}
\examples{
 \dontrun{
 launchApp()
 }

}
\author{
Rico Derks
}
