% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_to_env.R
\name{load_to_env}
\alias{load_to_env}
\title{Load to environment}
\usage{
load_to_env(RData, env = new.env())
}
\arguments{
\item{RData}{the .Rdata file.}

\item{env}{an environment to load the data into}
}
\value{
Returns an environment containing all the saved objects.
}
\description{
Load objects into a new environment
}
\details{
This function was found in a blog on R-bloggers. For more information
see the blog.
}
\references{
https://www.r-bloggers.com/2016/04/safe-loading-of-rdata-files/
}
\author{
Rico Derks
}
