% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{merge_data}
\alias{merge_data}
\title{Merge the meta data with the lipid data}
\usage{
merge_data(lipid_data, meta_data = NULL, by = NULL)
}
\arguments{
\item{lipid_data}{data frame containing the lipid data (wide format).}

\item{meta_data}{data frame containing the meta data.}

\item{by}{column name (character) of the the column to use to merge the meta
data with the lipid data}
}
\value{
Returns a merged data frame in tidy format
}
\description{
Merge the meta data with the lipid data.
}
\details{
If there is not meta data the lipid data will be converted to tidy format.
}
\author{
Rico Derks
}
