% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_explained_var_plot.R
\name{pca_explained_var_plot}
\alias{pca_explained_var_plot}
\title{Make an explained variance plot}
\usage{
pca_explained_var_plot(exp_var_data, num_pc = 5)
}
\arguments{
\item{exp_var_data}{tibble containing the scores information}

\item{num_pc}{number of components to plot, default is 5}
}
\value{
a plotly graph
}
\description{
Make an explained variance plot for PCA with plotly
}
\author{
Rico Derks
}
