% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_loadings_plot.R
\name{pca_loadings_plot}
\alias{pca_loadings_plot}
\title{Make a PCA loadings plot}
\usage{
pca_loadings_plot(loadings_data, xaxis = "PC1", yaxis = "PC2")
}
\arguments{
\item{loadings_data}{tibble containing the loadings information}

\item{xaxis}{which component to show on the x-axis (default = "PC1")}

\item{yaxis}{which component to show on the y-axis (default = "PC2")}
}
\value{
a plotly graph
}
\description{
Make a PCA loadings plot with plot_ly
}
\author{
Rico Derks
}
