% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_observation_plot.R
\name{pca_observation_plot}
\alias{pca_observation_plot}
\title{Make a PCA observation plot}
\usage{
pca_observation_plot(obs_data, var_name, color_by = "none")
}
\arguments{
\item{obs_data}{tibble containing the varaible information}

\item{var_name}{character vector of length 1 with the name of the variable}

\item{color_by}{color the observation by what}
}
\value{
a plotly graph
}
\description{
Make a PCA obaservation plot with plot_ly
}
\author{
Rico Derks
}
