% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_variable_plot.R
\name{pca_variable_plot}
\alias{pca_variable_plot}
\title{Make a PCA variable plot}
\usage{
pca_variable_plot(var_data, sample_name)
}
\arguments{
\item{var_data}{tibble containing the varaible information}

\item{sample_name}{character vector (length 1) with the sample name}
}
\value{
a plotly graph
}
\description{
Make a PCA variable plot with plot_ly
}
\author{
Rico Derks
}
