% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_msdial.R
\name{read_msdial}
\alias{read_msdial}
\title{Read MS-DIAl fiels}
\usage{
read_msdial(filename)
}
\arguments{
\item{filename}{The filename of the MS-DIAL file.}
}
\value{
Returns a data frame
}
\description{
Read the result files from MS-DIAL.
}
\author{
Rico Derks
}
