% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_lipid.R
\name{rename_lipid}
\alias{rename_lipid}
\title{Rename lipids automatically}
\usage{
rename_lipid(lipid_data, rename_info)
}
\arguments{
\item{lipid_data}{data frame containing the lipid data (wide format).}

\item{rename_info}{data frame containing the information which lipid to rename}
}
\value{
Returns a data frame with the lipid renamed.
}
\description{
Rename a seleceted lipid automatically. Append the lipid name with a number.
}
\details{
rename_info contains more then only which lipids to rename, ignore the rest.
}
\author{
Rico Derks
}
