% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_identified.R
\name{select_identified}
\alias{select_identified}
\title{Keep only the identified lipids}
\usage{
select_identified(df)
}
\arguments{
\item{df}{The tibble.}
}
\value{
Returns a tibble
}
\description{
Filter the tibble to keep only the identified lipids.
}
\author{
Rico Derks
}
