% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_issue_info.R
\name{set_issue_info}
\alias{set_issue_info}
\title{Merge the meta data with the lipid data}
\usage{
set_issue_info(lipid_data, info)
}
\arguments{
\item{lipid_data}{data frame containing the lipid data.}

\item{info}{data frame containing the issue info}
}
\value{
Returns the same data frame as lipid_data, but with the info added
}
\description{
Merge the meta data with the lipid data.
}
\details{
If there is not meta data the lipid data will be converted to tidy format.
}
\author{
Rico Derks
}
