% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyAppServer.R
\name{shinyAppServer}
\alias{shinyAppServer}
\title{Shiny app server function}
\usage{
shinyAppServer(input, output, session)
}
\arguments{
\item{input}{provided by shiny}

\item{output}{provided by shiny}

\item{session}{provided by shiny}
}
\description{
This is the server function to run the shiny app.
}
\author{
Rico Derks
}
