% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_lipids.R
\name{tidy_lipids}
\alias{tidy_lipids}
\title{Make the tibble in tidy format}
\usage{
tidy_lipids(df)
}
\arguments{
\item{df}{The tibble.}
}
\value{
Returns a tibble in tidy (long) format
}
\description{
Filter the tibble to keep only the identified lipids.
}
\details{
After making the tibble in long format also some additional columns ared added.
}
\author{
Rico Derks
}
