% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubblePlotServer.R
\name{bubblePlotServer}
\alias{bubblePlotServer}
\title{Create bubble plot server part}
\usage{
bubblePlotServer(id, lipid_data, pattern, title)
}
\arguments{
\item{id}{input id}

\item{lipid_data}{data which is used to make the plot.}

\item{pattern}{regular expression pattern to select the correct lipid classes.}

\item{title}{is the title use on top of the indentication page}
}
\value{
a part of the server
}
\description{
Create the server part for the bubble plot
}
\author{
Rico Derks
}
