% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubblePlotServer.R
\name{bubblePlotServer}
\alias{bubblePlotServer}
\title{Create bubble plot server part}
\usage{
bubblePlotServer(id, data, pattern, lipid_data)
}
\arguments{
\item{id}{input id}

\item{data}{data which is used to make the plot.}

\item{pattern}{regular expression pattern to select the correct lipid classes.}

\item{lipid_data}{the wide data frame.}
}
\value{
a part of the server
}
\description{
Create the server part for the bubble plot
}
\author{
Rico Derks
}
