% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubblePlotUI.R
\name{bubblePlotUI}
\alias{bubblePlotUI}
\title{Create bubble plot UI}
\usage{
bubblePlotUI(id)
}
\arguments{
\item{id}{id of the UI}
}
\value{
a part of the UI
}
\description{
Create the UI for the bubble plot
}
\author{
Rico Derks
}
