% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_blank_ratio.R
\name{calc_blank_ratio}
\alias{calc_blank_ratio}
\title{Calculate the sample to average blank ratio}
\usage{
calc_blank_ratio(lipid_data, ratio = 5)
}
\arguments{
\item{lipid_data}{tibble in tidy format}

\item{ratio}{numeric(1) cutoff ratio of sample/ blank}
}
\value{
Returns a tibble in tidy with sample / blank ratio for each lipid
    species.
}
\description{
Calculate for each lipid the ratio between the sample and the
    average of all blanks.
}
\author{
Rico Derks
}
