% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_heatmap.R
\name{cor_heatmap}
\alias{cor_heatmap}
\title{Create correlation heatmap of all samples}
\usage{
cor_heatmap(df)
}
\arguments{
\item{df}{tibble in tidy format}
}
\value{
pheatmap object
}
\description{
Create a correlation heatmap of all samples..
}
\author{
Rico Derks
}
