% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_heatmap2.R
\name{cor_heatmap2}
\alias{cor_heatmap2}
\title{Create correlation heatmap of all samples}
\usage{
cor_heatmap2(lipid_data)
}
\arguments{
\item{lipid_data}{tibble with all the lipid data}
}
\value{
plotly object
}
\description{
Create a correlation heatmap of all samples..
}
\author{
Rico Derks
}
