% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_scores_plot.R
\name{pca_scores_plot}
\alias{pca_scores_plot}
\title{Make a PCA scores plot}
\usage{
pca_scores_plot(scores_data, xaxis = "PC1", yaxis = "PC2")
}
\arguments{
\item{scores_data}{tibble containing the scores information}

\item{xaxis}{which component to show on the x-axis (default = "PC1")}

\item{yaxis}{which component to show on the y-axis (default = "PC2")}
}
\value{
a plotly graph
}
\description{
Make a PCA scores plot with plot_ly
}
\author{
Rico Derks
}
