% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_rsd_lipidclass_violin.R
\name{show_rsd_lipidclass_violin}
\alias{show_rsd_lipidclass_violin}
\title{Show a violin plot of all RSD values per lipid class}
\usage{
show_rsd_lipidclass_violin(qc_data)
}
\arguments{
\item{qc_data}{tibble in tidy format}
}
\value{
ggplot2 object
}
\description{
Show a violin plot of all RSD values of all lipids per lipid class
    within all QCpool samples..
}
\author{
Rico Derks
}
