% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_plot.R
\name{volcano_plot}
\alias{volcano_plot}
\title{Create volcano plot}
\usage{
volcano_plot(lipid_data, pvalue_adjust = FALSE, title = "")
}
\arguments{
\item{lipid_data}{tibble with all the lipid data and test data}

\item{pvalue_adjust}{show the corrected p value, default is FALSE}

\item{title}{title of the plot}
}
\value{
plotly object
}
\description{
Create volcano plot.
}
\author{
Rico Derks
}
